Working on synchronizer

+++ src/main/java/org/sikrip/vboeditor/gui/GPSViewer.java
@@ -132,7 +132,7 @@ final class GPSViewer extends JPanel implements ActionListener {
         playPause.setEnabled(enable);
     }
 
-    double getCurrentTime() {
+    long getCurrentTime() {
         return currentPositionIdx * interval;
     }
